requirements.txt
langgraph
langchain-anthropic
langchain-openai
langchain-google-genai
langchain-community
faiss-cpu
pypdf
python-dotenv
rich
streamlit
